Add check of appropriate status in `GetMaterialDeleteConfirmationDataAction`

+++ api/src/org/labkey/api/exp/api/ExpMaterial.java
@@ -45,5 +45,5 @@ public interface ExpMaterial extends ExpRunItem
 
     String getAliquotedFromLSID();
 
-    Long getStatus();
+    Integer getStatus();
 }
